changed rating to a double and addded error checking mechanism to more inputs

+++ coursemaker/src/TupleComparator.java
@@ -1,9 +1,9 @@
 import java.util.Comparator;
 
-public class TupleComparator implements Comparator<Tuple<Integer, Course>> {
+public class TupleComparator implements Comparator<Tuple<Double, Course>> {
 //    class for Tuple Comparator
     @Override
-    public int compare(Tuple<Integer, Course> t1, Tuple<Integer, Course> t2) {
+    public int compare(Tuple<Double, Course> t1, Tuple<Double, Course> t2) {
         return t1.getFirst().compareTo(t2.getFirst());
     }
 }
